+++ server/controller/messageCenter/addMessage.ts
@@ -1,7 +1,7 @@
 import { BaseProtocol, DB, Schemas } from '@server/db'
 import { ActivityLogRepository } from '@server/repository/assessment/activityLog'
-import { MessageRepository } from '@server/repository/public/message'
-import { MessageTopicRepository } from '@server/repository/public/messageTopic'
+import { MessageRepository } from '@server/repository/assessmentCycle/message'
+import { MessageTopicRepository } from '@server/repository/assessmentCycle/messageTopic'
 
 import { CountryIso } from '@meta/area'
 import { ActivityLogMessage, Assessment, Cycle } from '@meta/assessment'
